feat: Add function to summarize transactions
This commit adds a new function `resGetSummarizeTransactions` to the `main.go` file. This function generates dummy transactions and sends them to the `GetSummarizeTransactions` method of the `BankAdapter` struct. The generated transactions include the transaction amount, type (IN or OUT), and notes. The function then receives the transaction summary from the server and logs it.

The purpose of this change is to provide a way to summarize transactions for a given account. This feature will be used to display a summary of recent transactions to the user.